Cast the sky level to plain float before publishing
get_sky_level returned numpy.float64, which msgspec cannot encode, so the
expose_complete event was silently dropped for every subscriber
(2026-07-21 19:23: TypeError: Encoding objects of type numpy.float64 is
unsupported). robobs now counts flat frames from this event, so a dropped
publication means the skyflat ledger records nothing.

+++ src/chimera_skyflat/controllers/autoskyflat.py
@@ -379,7 +379,9 @@ class AutoSkyFlat(ChimeraObject, IAutoSkyFlat):
         """
 
         frame = fits.getdata(filename)
-        img_mean = np.mean(frame)
+        # plain float: numpy.float64 is not msgspec-serializable, so leaking
+        # it into the expose_complete event silently drops the publication
+        img_mean = float(np.mean(frame))
         return img_mean
 
     def abort(self):
